requirements.txt
streamlit==1.10.0
Pillow==8.4.0
pytesseract==0.3.8
